# openapi.yaml
openapi: 3.0.3
info:
  title: Google Drive API
  description: An API to interact with Google Drive, allowing users to manage files, folders, and presentations.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /drive/list/{path}:
    get:
      summary: List Files in a Folder
      description: Retrieves the contents of a specific folder in Google Drive by its full path.
      operationId: list_drive_path_drive_list__path__get
      parameters:
        - name: path
          in: path
          required: true
          description: The full path of the folder to list the contents of.
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileItem'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drive/search:
    get:
      summary: Search Files and Folders
      description: Searches for files and folders by name in Google Drive.
      operationId: search_drive_drive_search_get
      parameters:
        - name: name
          in: query
          required: false
          description: The name of the file or folder to search for.
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResultItem'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drive/documents:
    post:
      summary: Create Google Document
      description: Creates a new, empty Google Document.
      operationId: create_document_drive_documents_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '201':
          description: Document Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInfo'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drive/documents/{document_id}:
    get:
      summary: Get Google Document
      description: Retrieves a Google Document by its ID in a simplified format.
      operationId: get_document_drive_documents__document_id__get
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /drive/documents/{document_id}:append:
    post:
      summary: Append to Document
      description: Appends text to an existing Google Document.
      operationId: append_to_document_drive_documents__document_id___append_post
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendTextPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFoundError'
  /drive/spreadsheets:
    post:
      summary: Create Google Spreadsheet
      description: Creates a new, empty Google Spreadsheet.
      operationId: create_spreadsheet_drive_spreadsheets_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpreadsheetCreate'
      responses:
        '201':
          description: Spreadsheet Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadsheetInfo'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drive/spreadsheets/{spreadsheet_id}:
    get:
      summary: Get Google Spreadsheet
      description: Retrieves a Google Spreadsheet by its ID, including all sheets and their values.
      operationId: get_spreadsheet_drive_spreadsheets__spreadsheet_id__get
      parameters:
        - name: spreadsheet_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spreadsheet'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /drive/spreadsheets/{spreadsheet_id}/sheets:
    post:
      summary: Add Sheet
      description: Adds a new sheet to an existing spreadsheet.
      operationId: add_sheet_drive_spreadsheets__spreadsheet_id__sheets_post
      parameters:
        - name: spreadsheet_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSheetPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
  /drive/spreadsheets/{spreadsheet_id}/sheets/{sheet_name}:append:
    post:
      summary: Append Data to Sheet
      description: Appends data to a specific sheet in a spreadsheet.
      operationId: append_data_drive_spreadsheets__spreadsheet_id__sheets__sheet_name___append_post
      parameters:
        - name: spreadsheet_id
          in: path
          required: true
          schema:
            type: string
        - name: sheet_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendDataPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
  /drive/spreadsheets/{spreadsheet_id}/sheets/{sheet_name}:clear:
    post:
      summary: Clear Sheet
      description: Clears all data from a specific sheet in a spreadsheet.
      operationId: clear_sheet_drive_spreadsheets__spreadsheet_id__sheets__sheet_name___clear_post
      parameters:
        - name: spreadsheet_id
          in: path
          required: true
          schema:
            type: string
        - name: sheet_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
  /drive/presentations:
    post:
      summary: Create Google Presentation
      description: Creates a new, empty Google Slides presentation.
      operationId: create_google_presentation_drive_presentations_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationCreate'
      responses:
        '201':
          description: Presentation Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationInfo'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drive/presentations/{presentation_id}:
    get:
      summary: Get Google Presentation
      description: Retrieves a Google Slides presentation by its ID.
      operationId: get_google_presentation_drive_presentations__presentation_id__get
      parameters:
        - name: presentation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleSlides'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /drive/presentations/{presentation_id}/slides:
    post:
      summary: Add Slide to Presentation
      description: Adds a new slide with a specified layout and content to an existing presentation.
      operationId: add_slide_to_presentation_drive_presentations__presentation_id__slides_post
      parameters:
        - name: presentation_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlideContent'
      responses:
        '201':
          description: Slide Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlideCreationResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    FileItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
    SearchResultItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        path:
          type: string
    Error:
      type: object
      properties:
        detail:
          type: string
    Document:
      type: object
      properties:
        document_id:
          type: string
        title:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/Paragraph'
    Paragraph:
      type: object
      properties:
        text:
          type: string
    DocumentCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The title for the new document.
    DocumentInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    AppendTextPayload:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: The text to append to the document.
    GoogleDocument:
      type: object
      description: The JSON representation of a Google Document.
    Spreadsheet:
      type: object
      properties:
        spreadsheet_id:
          type: string
        title:
          type: string
        sheets:
          type: array
          items:
            $ref: '#/components/schemas/Sheet'
    Sheet:
      type: object
      properties:
        title:
          type: string
        values:
          type: array
          items:
            type: array
            items:
              type: object
    SpreadsheetCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
    SpreadsheetInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    AppendDataPayload:
      type: object
      required:
        - values
      properties:
        values:
          type: array
          items:
            type: array
            items:
              type: object
    AddSheetPayload:
      type: object
      required:
        - title
      properties:
        title:
          type: string
    GoogleSpreadsheet:
      type: object
      description: The JSON representation of a Google Spreadsheet.
    GoogleSlides:
      type: object
      description: The JSON representation of a Google Slides presentation.
    Presentation:
      type: object
      properties:
        presentation_id:
          type: string
        title:
          type: string
        slides:
          type: array
          items:
            $ref: '#/components/schemas/Slide'
    Slide:
      type: object
      properties:
        slide_id:
          type: string
        layout:
          type: string
        title:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
    PresentationCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The title for the new presentation.
          example: "My New Presentation"
    PresentationInfo:
      type: object
      properties:
        id:
          type: string
          description: The ID of the newly created presentation.
        name:
          type: string
          description: The name of the newly created presentation.
    SlidePayload:
      type: object
      required:
        - layout
      properties:
        layout:
          type: string
          description: The layout for the new slide.
          example: "TITLE_AND_BODY"
        title:
          type: string
          description: The title text for the slide.
        subtitle:
          type: string
          description: The subtitle text for the slide.
        body:
          type: string
          description: The body text for the slide.
    SlideCreationResponse:
      type: object
      properties:
        slideId:
          type: string
          description: The ID of the newly created slide.
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'